requirements.txt
pandas
streamlit
matplotlib.pyplot
prettify
